+++ backend/canisters/community/impl/src/lib.rs
@@ -11,7 +11,7 @@ use fire_and_forget_handler::FireAndForgetHandler;
 use group_chat_core::AccessRulesInternal;
 use model::{events::CommunityEvents, invited_users::InvitedUsers, members::CommunityMemberInternal};
 use notifications_canister::c2c_push_notification;
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use serde_bytes::ByteBuf;
 use std::cell::RefCell;
 use std::ops::Deref;
@@ -24,6 +24,7 @@ use utils::env::Environment;
 use utils::regular_jobs::RegularJobs;
 
 mod activity_notifications;
+mod data_deserialize;
 mod guards;
 mod jobs;
 mod lifecycle;
@@ -169,7 +170,7 @@ impl RuntimeState {
     }
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize)]
 struct Data {
     is_public: bool,
     name: String,
@@ -197,8 +198,11 @@ struct Data {
     timer_jobs: TimerJobs<TimerJob>,
     fire_and_forget_handler: FireAndForgetHandler,
     activity_notification_state: ActivityNotificationState,
+    #[serde(default)]
     groups_being_imported: GroupsBeingImported,
+    #[serde(default)]
     test_mode: bool,
+    #[serde(default)]
     cached_chat_metrics: Timestamped<ChatMetrics>,
 }