+++ backend/canisters/community/CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Extend versioned rules to communities and groups ([#4219](https://github.com/open-chat-labs/open-chat/pull/4219))
 - Make importing groups more efficient ([#4239](https://github.com/open-chat-labs/open-chat/pull/4239))
+- Hack to allow upgrading the OpenChat community without exceeding the instruction limit ([#4245](https://github.com/open-chat-labs/open-chat/pull/4245))
 
 ### Fixed